+++ iroh/src/node/protocol/gossip.rs
@@ -1,0 +1,13 @@
+use std::sync::Arc;
+
+use anyhow::Result;
+use futures_lite::future::Boxed as BoxedFuture;
+use iroh_net::endpoint::Connecting;
+
+use super::ProtocolHandler;
+
+impl ProtocolHandler for iroh_gossip::net::Gossip {
+    fn accept(self: Arc<Self>, conn: Connecting) -> BoxedFuture<Result<()>> {
+        Box::pin(async move { self.handle_connection(conn.await?).await })
+    }
+}
